@@ -1,9 +1,9 @@
 [package]
 name = "czkawka_core"
-version = "5.0.2"
+version = "5.1.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.65.0"
 description = "Core of Czkawka app"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
@@ -11,52 +11,52 @@ repository = "https://github.com/qarmin/czkawka"
 
 
 [dependencies]
-humansize = "1.1.1"
-rayon = "1.5.3"
+humansize = "2.1.3"
+rayon = "1.6.1"
 crossbeam-channel = "0.5.6"
 
 # For saving/loading config files to specific directories
 directories-next = "2.0.0"
 
 # Needed by similar images
-image_hasher = "1.0.0"
+image_hasher = "1.1.2"
 bk-tree = "0.4.0"
-image = "0.24.3"
+image = "0.24.5"
 hamming = "0.1.3"
 
 # Needed by same music
 bitflags = "1.3.2"
-lofty = "0.8.0"
+lofty = "0.11.0"
 
 # Futures - needed by async progress sender
-futures = "0.3.23"
+futures = "0.3.26"
 
 # Needed by broken files
-zip = { version = "0.6.2", features = ["aes-crypto", "bzip2", "deflate", "time"], default-features = false }
+zip = { version = "0.6.4", features = ["aes-crypto", "bzip2", "deflate", "time"], default-features = false }
 audio_checker = "0.1.0"
-pdf = "0.7.2"
+pdf = "0.8.0"
 
 # Hashes for duplicate files
-blake3 = "1.3.1"
+blake3 = "1.3.3"
 crc32fast = "1.3.2"
-xxhash-rust = { version = "0.8.5", features = ["xxh3"] }
+xxhash-rust = { version = "0.8.6", features = ["xxh3"] }
 
 tempfile = "3.3.0"
 
-# Video Duplactes
+# Video Duplicates
 vid_dup_finder_lib = "0.1.1"
 ffmpeg_cmdline_utils = "0.1.2"
 
 # Saving/Loading Cache
-serde = "1.0.143"
+serde = "1.0"
 bincode = "1.3.3"
-serde_json = "1.0.83"
+serde_json = "1.0"
 
 # Language
-i18n-embed = { version = "0.13.4", features = ["fluent-system", "desktop-requester"] }
-i18n-embed-fl = "0.6.4"
-rust-embed = "6.4.0"
-once_cell = "1.13.1"
+i18n-embed = { version = "0.13.8", features = ["fluent-system", "desktop-requester"] }
+i18n-embed-fl = "0.6.5"
+rust-embed = "6.4.2"
+once_cell = "1.17.1"
 
 # Raw image files
 rawloader = "0.37.1"
@@ -64,13 +64,15 @@ imagepipe = "0.5.0"
 
 # Checking for invalid extensions
 mime_guess = "2.0.4"
-infer = "0.9.0"
+infer = "0.12.0"
 
-num_cpus = "1.13.1"
+num_cpus = "1.15.0"
 
 # Heif/Heic
-libheif-rs = { version = "0.15.0", optional = true }
-anyhow = { version = "1.0.62", optional = true }
+libheif-rs = { version = "0.16.0", optional = true }
+anyhow = { version = "1.0", optional = true }
+
+state="0.5.3"
 
 [features]
 default = []
