@@ -1,65 +1,65 @@
 [package]
 name = "czkawka_gui"
-version = "5.0.2"
+version = "5.1.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.65.0"
 description = "GTK frontend of Czkawka"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
 repository = "https://github.com/qarmin/czkawka"
 
 [dependencies]
-gdk4 = "0.4.8"
-glib = "0.15.12"
+gdk4 = "0.6.0"
+glib = "0.17.1"
 
-humansize = "1.1.1"
-chrono = "0.4.22"
+humansize = "2.1.3"
+chrono = "0.4.23"
 
 # Used for sending stop signal across threads
 crossbeam-channel = "0.5.6"
 
-# To get informations about progress
-futures = "0.3.23"
+# To get information about progress
+futures = "0.3.26"
 
 # For saving/loading config files to specific directories
 directories-next = "2.0.0"
 
 # For opening files
-open = "3.0.2"
+open = "3.2.0"
 
 # To get image preview
-image = "0.24.3"
+image = "0.24.5"
 
 # To be able to use custom select
-regex = "1.6.0"
+regex = "1.7.1"
 
 # To get image_hasher types
-image_hasher = "1.0.0"
+image_hasher = "1.1.2"
 
 # Move files to trash
-trash = "2.1.5"
+trash = "3.0.1"
 
 # For moving files(why std::fs doesn't have such features)
-fs_extra = "1.2.0"
+fs_extra = "1.3.0"
 
 # Language
-i18n-embed = { version = "0.13.4", features = ["fluent-system", "desktop-requester"] }
-i18n-embed-fl = "0.6.4"
-rust-embed = "6.4.0"
-once_cell = "1.13.1"
+i18n-embed = { version = "0.13.8", features = ["fluent-system", "desktop-requester"] }
+i18n-embed-fl = "0.6.5"
+rust-embed = "6.4.2"
+once_cell = "1.17.1"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3.9", features = ["combaseapi", "objbase", "shobjidl_core", "windef", "winerror", "wtypesbase", "winuser"] }
 
 [dependencies.gtk4]
-version = "0.4.8"
+version = "0.6.1"
 default-features = false
 features = ["v4_6"]
 
 [dependencies.czkawka_core]
 path = "../czkawka_core"
-version = "5.0.2"
+version = "5.1.0"
 features = []
 
 [features]
