@@ -3,9 +3,10 @@ use std::collections::HashMap;
 use std::path::PathBuf;
 
 use gdk4::gdk_pixbuf::{InterpType, Pixbuf};
+use glib::signal::Inhibit;
 use glib::Error;
 use gtk4::prelude::*;
-use gtk4::{ListStore, TextView, TreeView, Widget};
+use gtk4::{ListStore, ScrollType, TextView, TreeView, Widget};
 use image::codecs::jpeg::JpegEncoder;
 use image::{DynamicImage, EncodableLayout};
 use once_cell::sync::OnceCell;
@@ -44,7 +45,7 @@ pub const KEY_SPACE: u32 = 65;
 // pub const KEY_HOME: u32 = 115;
 // pub const KEY_END: u32 = 110;
 
-#[derive(Eq, PartialEq)]
+#[derive(Eq, PartialEq, Copy, Clone, Debug)]
 pub enum PopoverTypes {
     All,
     Size,
@@ -63,6 +64,7 @@ pub enum BottomButtonsEnum {
     Hardlink,
     Move,
     Compare,
+    Sort,
 }
 
 pub enum Message {
@@ -79,11 +81,13 @@ pub enum Message {
     BadExtensions(BadExtensions),
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsDuplicates {
     // Columns for duplicate treeview
     ActivatableSelectButton = 0,
     SelectionButton,
     Size,
+    SizeAsBytes,
     Name,
     Path,
     Modification,
@@ -93,6 +97,7 @@ pub enum ColumnsDuplicates {
     TextColor,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsEmptyFolders {
     // Columns for empty folder treeview
     SelectionButton = 0,
@@ -102,17 +107,20 @@ pub enum ColumnsEmptyFolders {
     ModificationAsSecs,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsIncludedDirectory {
     // Columns for Included Directories in upper Notebook
     Path = 0,
     ReferenceButton,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsExcludedDirectory {
     // Columns for Excluded Directories in upper Notebook
     Path = 0,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsBigFiles {
     SelectionButton = 0,
     Size,
@@ -123,6 +131,7 @@ pub enum ColumnsBigFiles {
     ModificationAsSecs,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsEmptyFiles {
     SelectionButton = 0,
     Name,
@@ -131,6 +140,7 @@ pub enum ColumnsEmptyFiles {
     ModificationAsSecs,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsTemporaryFiles {
     SelectionButton = 0,
     Name,
@@ -139,6 +149,7 @@ pub enum ColumnsTemporaryFiles {
     ModificationAsSecs,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsSimilarImages {
     ActivatableSelectButton = 0,
     SelectionButton,
@@ -155,6 +166,7 @@ pub enum ColumnsSimilarImages {
     TextColor,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsSimilarVideos {
     ActivatableSelectButton = 0,
     SelectionButton,
@@ -169,6 +181,7 @@ pub enum ColumnsSimilarVideos {
     TextColor,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsSameMusic {
     ActivatableSelectButton = 0,
     SelectionButton,
@@ -190,6 +203,7 @@ pub enum ColumnsSameMusic {
     TextColor,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsInvalidSymlinks {
     SelectionButton = 0,
     Name,
@@ -200,6 +214,7 @@ pub enum ColumnsInvalidSymlinks {
     ModificationAsSecs,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsBrokenFiles {
     SelectionButton = 0,
     Name,
@@ -209,6 +224,7 @@ pub enum ColumnsBrokenFiles {
     ModificationAsSecs,
 }
 
+#[derive(Clone, Copy)]
 pub enum ColumnsBadExtensions {
     SelectionButton = 0,
     Name,
@@ -252,12 +268,12 @@ pub fn get_string_from_list_store(tree_view: &TreeView, column_full_path: i32, c
     }
 }
 
-pub fn get_path_buf_from_vector_of_strings(vec_string: Vec<String>) -> Vec<PathBuf> {
+pub fn get_path_buf_from_vector_of_strings(vec_string: &[String]) -> Vec<PathBuf> {
     vec_string.iter().map(PathBuf::from).collect()
 }
 
 pub fn print_text_messages_to_text_view(text_messages: &Messages, text_view: &TextView) {
-    let mut messages: String = String::from("");
+    let mut messages: String = String::new();
     if !text_messages.messages.is_empty() {
         messages += format!("############### {}({}) ###############\n", flg!("text_view_messages"), text_messages.messages.len()).as_str();
     }
@@ -302,7 +318,7 @@ pub fn add_text_to_text_view(text_view: &TextView, string_to_append: &str) {
     if current_text.is_empty() {
         buffer.set_text(string_to_append);
     } else {
-        buffer.set_text(format!("{}\n{}", current_text, string_to_append).as_str());
+        buffer.set_text(format!("{current_text}\n{string_to_append}").as_str());
     }
 }
 
@@ -337,7 +353,7 @@ pub fn get_dialog_box_child(dialog: &gtk4::Dialog) -> gtk4::Box {
     dialog.child().unwrap().downcast::<gtk4::Box>().unwrap()
 }
 
-pub fn change_dimension_to_krotka(dimensions: String) -> (u64, u64) {
+pub fn change_dimension_to_krotka(dimensions: &str) -> (u64, u64) {
     #[allow(clippy::single_char_pattern)]
     let vec = dimensions.split::<&str>("x").collect::<Vec<_>>();
     assert_eq!(vec.len(), 2); // 400x400 - should only have two elements, if have more, then something is not good
@@ -394,9 +410,8 @@ pub fn clean_invalid_headers(model: &ListStore, column_header: i32, column_path:
     if let Some(first_iter) = model.iter_first() {
         let mut vec_tree_path_to_delete: Vec<gtk4::TreePath> = Vec::new();
         let mut current_iter = first_iter;
-        if !model.get::<bool>(&current_iter, column_header) {
-            panic!("First deleted element, should be a header"); // First element should be header
-        };
+        // First element should be header
+        assert!(model.get::<bool>(&current_iter, column_header), "First deleted element, should be a header");
 
         let mut next_iter;
         let mut next_next_iter;
@@ -404,9 +419,8 @@ pub fn clean_invalid_headers(model: &ListStore, column_header: i32, column_path:
         // Empty means default check type
         if model.get::<String>(&current_iter, column_path).is_empty() {
             'main: loop {
-                if !model.get::<bool>(&current_iter, column_header) {
-                    panic!("First deleted element, should be a header"); // First element should be header
-                };
+                // First element should be header
+                assert!(model.get::<bool>(&current_iter, column_header), "First deleted element, should be a header");
 
                 next_iter = current_iter;
                 if !model.iter_next(&next_iter) {
@@ -457,9 +471,8 @@ pub fn clean_invalid_headers(model: &ListStore, column_header: i32, column_path:
         // Non empty means that header points at reference folder
         else {
             'reference: loop {
-                if !model.get::<bool>(&current_iter, column_header) {
-                    panic!("First deleted element, should be a header"); // First element should be header
-                };
+                // First element should be header
+                assert!(model.get::<bool>(&current_iter, column_header), "First deleted element, should be a header");
 
                 next_iter = current_iter;
                 if !model.iter_next(&next_iter) {
@@ -586,7 +599,7 @@ pub fn count_number_of_groups(tree_view: &TreeView, column_header: i32) -> u32 {
     number_of_selected_groups
 }
 
-pub fn resize_pixbuf_dimension(pixbuf: Pixbuf, requested_size: (i32, i32), interp_type: InterpType) -> Option<Pixbuf> {
+pub fn resize_pixbuf_dimension(pixbuf: &Pixbuf, requested_size: (i32, i32), interp_type: InterpType) -> Option<Pixbuf> {
     let current_ratio = pixbuf.width() as f32 / pixbuf.height() as f32;
     let mut new_size;
     match current_ratio.partial_cmp(&(requested_size.0 as f32 / requested_size.1 as f32)).unwrap() {
@@ -613,8 +626,8 @@ pub fn get_max_file_name(file_name: &str, max_length: usize) -> String {
         let start_characters = 10;
         let difference = characters_in_filename - max_length;
         let second_part_start = start_characters + difference;
-        let mut string_pre = "".to_string();
-        let mut string_after = "".to_string();
+        let mut string_pre = String::new();
+        let mut string_after = String::new();
 
         for (index, character) in file_name.chars().enumerate() {
             if index < start_characters {
@@ -636,9 +649,8 @@ pub fn get_custom_label_from_widget<P: IsA<Widget>>(item: &P) -> gtk4::Label {
     while let Some(widget) = widgets_to_check.pop() {
         if let Ok(label) = widget.clone().downcast::<gtk4::Label>() {
             return label;
-        } else {
-            widgets_to_check.extend(get_all_direct_children(&widget));
         }
+        widgets_to_check.extend(get_all_direct_children(&widget));
     }
     panic!("Button doesn't have proper custom label child");
 }
@@ -649,9 +661,8 @@ pub fn get_custom_image_from_widget<P: IsA<Widget>>(item: &P) -> gtk4::Image {
     while let Some(widget) = widgets_to_check.pop() {
         if let Ok(image) = widget.clone().downcast::<gtk4::Image>() {
             return image;
-        } else {
-            widgets_to_check.extend(get_all_direct_children(&widget));
         }
+        widgets_to_check.extend(get_all_direct_children(&widget));
     }
     panic!("Button doesn't have proper custom label child");
 }
@@ -668,7 +679,7 @@ pub fn debug_print_widget<P: IsA<Widget>>(item: &P) {
             widgets_to_check.push((next_free_number, current_number, widget));
             next_free_number += 1;
         }
-        println!("{}, {}, {:?} ", current_number, parent_number, widget);
+        println!("{current_number}, {parent_number}, {widget:?} ");
     }
 }
 
@@ -728,7 +739,7 @@ pub fn get_pixbuf_from_dynamic_image(dynamic_image: &DynamicImage) -> Result<Pix
 pub fn check_if_value_is_in_list_store(list_store: &ListStore, column: i32, value: &str) -> bool {
     if let Some(iter) = list_store.iter_first() {
         loop {
-            let list_store_value: String = list_store.get::<String>(&iter, column as i32);
+            let list_store_value: String = list_store.get::<String>(&iter, column);
 
             if value == list_store_value {
                 return true;
@@ -746,7 +757,7 @@ pub fn check_if_value_is_in_list_store(list_store: &ListStore, column: i32, valu
 pub fn check_if_list_store_column_have_all_same_values(list_store: &ListStore, column: i32, value: bool) -> bool {
     if let Some(iter) = list_store.iter_first() {
         loop {
-            let list_store_value: bool = list_store.get::<bool>(&iter, column as i32);
+            let list_store_value: bool = list_store.get::<bool>(&iter, column);
 
             if value != list_store_value {
                 return false;
@@ -759,6 +770,13 @@ pub fn check_if_list_store_column_have_all_same_values(list_store: &ListStore, c
         return true;
     }
     false
+}
+
+pub fn scale_step_function(scale: &gtk4::Scale, _scroll_type: ScrollType, value: f64) -> Inhibit {
+    scale.set_increments(1_f64, 1_f64);
+    scale.set_round_digits(0);
+    scale.set_fill_level(value.round());
+    Inhibit(false)
 }
 
 #[cfg(test)]
@@ -864,8 +882,8 @@ mod test {
 
     #[test]
     fn test_change_dimension_to_krotka() {
-        assert_eq!(change_dimension_to_krotka("50x50".to_string()), (50, 50));
-        assert_eq!(change_dimension_to_krotka("6000x6000".to_string()), (6000, 6000));
+        assert_eq!(change_dimension_to_krotka("50x50"), (50, 50));
+        assert_eq!(change_dimension_to_krotka("6000x6000"), (6000, 6000));
     }
 
     #[gtk4::test]
