@@ -5,9 +5,10 @@ use std::io::{Read, Write};
 use std::path::{Path, PathBuf};
 use std::{env, fs};
 
+use czkawka_core::common::get_default_number_of_threads;
 use directories_next::ProjectDirs;
 use gtk4::prelude::*;
-use gtk4::{ComboBoxText, ScrolledWindow, TextView};
+use gtk4::{ComboBoxText, ScrolledWindow, TextView, TreeView};
 
 use czkawka_core::common_dir_traversal::CheckingMethod;
 use czkawka_core::similar_images::SIMILAR_VALUES;
@@ -48,6 +49,8 @@ const DEFAULT_BROKEN_FILES_AUDIO: bool = true;
 const DEFAULT_BROKEN_FILES_ARCHIVE: bool = true;
 const DEFAULT_BROKEN_FILES_IMAGE: bool = true;
 
+const DEFAULT_THREAD_NUMBER: u32 = 0;
+
 const DEFAULT_NUMBER_OF_BIGGEST_FILES: &str = "50";
 const DEFAULT_SIMILAR_IMAGES_SIMILARITY: f32 = 0.0;
 const DEFAULT_SIMILAR_IMAGES_IGNORE_SAME_SIZE: bool = false;
@@ -80,10 +83,10 @@ impl LoadSaveStruct {
         }
     }
 
-    pub fn get_vector_string(&self, key: String, default_value: Vec<String>) -> Vec<String> {
-        if self.loaded_items.contains_key(&key) {
+    pub fn get_vector_string(&self, key: &str, default_value: Vec<String>) -> Vec<String> {
+        if self.loaded_items.contains_key(key) {
             let mut new_vector = Vec::new();
-            for i in self.loaded_items.get(&key).unwrap() {
+            for i in self.loaded_items.get(key).unwrap() {
                 if !i.trim().is_empty() {
                     new_vector.push(i.trim().to_string());
                 }
@@ -95,10 +98,9 @@ impl LoadSaveStruct {
     }
     pub fn get_integer_string(&self, key: String, default_value: String) -> String {
         if default_value.parse::<i64>().is_err() {
-            println!("Default value {} can't be convert to integer value", default_value);
+            println!("Default value {default_value} can't be convert to integer value");
             panic!();
         }
-        assert!(default_value.parse::<i64>().is_ok());
         let mut returned_value = self.get_string(key, default_value.clone());
         if returned_value.parse::<i64>().is_err() {
             returned_value = default_value;
@@ -111,13 +113,13 @@ impl LoadSaveStruct {
             return if item.len() == 1 {
                 item[0].clone()
             } else if item.is_empty() {
-                "".to_string()
+                String::new()
             } else {
                 add_text_to_text_view(
                     &self.text_view,
                     &flg!(
                         "saving_loading_invalid_string",
-                        generate_translation_hashmap(vec![("key", key), ("result", format!("{:?}", item))])
+                        generate_translation_hashmap(vec![("key", key), ("result", format!("{item:?}"))])
                     ),
                 );
                 default_value
@@ -131,19 +133,18 @@ impl LoadSaveStruct {
             let item = self.loaded_items.get(&key).unwrap().clone().into_iter().filter(|e| !e.is_empty()).collect::<Vec<String>>();
 
             return if item.len() == 1 {
-                match item[0].parse::<T>() {
-                    Ok(t) => t,
-                    Err(_) => {
-                        println!("Failed to decode integer from \"{}\", found {:?}", key, item[0]);
-                        default_value
-                    }
+                if let Ok(t) = item[0].parse::<T>() {
+                    t
+                } else {
+                    println!("Failed to decode integer from \"{}\", found {:?}", key, item[0]);
+                    default_value
                 }
             } else {
                 add_text_to_text_view(
                     &self.text_view,
                     &flg!(
                         "saving_loading_invalid_int",
-                        generate_translation_hashmap(vec![("key", key), ("result", format!("{:?}", item))])
+                        generate_translation_hashmap(vec![("key", key), ("result", format!("{item:?}"))])
                     ),
                 );
                 default_value
@@ -156,7 +157,7 @@ impl LoadSaveStruct {
         if self.loaded_items.contains_key(&key) {
             let item = self.loaded_items.get(&key).unwrap().clone().into_iter().filter(|e| !e.is_empty()).collect::<Vec<String>>();
             return if item.len() == 1 {
-                let text = item[0].clone().trim().to_lowercase();
+                let text = item[0].trim().to_lowercase();
                 if text == "false" || text == "0" {
                     false
                 } else if text == "true" || text == "1" {
@@ -176,7 +177,7 @@ impl LoadSaveStruct {
                     &self.text_view,
                     &flg!(
                         "saving_loading_invalid_bool",
-                        generate_translation_hashmap(vec![("key", key), ("result", format!("{:?}", item))])
+                        generate_translation_hashmap(vec![("key", key), ("result", format!("{item:?}"))])
                     ),
                 );
                 default_value
@@ -187,7 +188,7 @@ impl LoadSaveStruct {
     }
 
     // Bool, int, string
-    pub fn save_var<T: ToString>(&mut self, key: String, value: T) {
+    pub fn save_var<T: ToString>(&mut self, key: String, value: &T) {
         if self.loaded_items.contains_key(&key) {
             add_text_to_text_view(
                 &self.text_view,
@@ -198,7 +199,7 @@ impl LoadSaveStruct {
         self.loaded_items.insert(key, vec![value.to_string()]);
     }
 
-    pub fn save_list_store(&mut self, key: String, tree_view: &gtk4::TreeView, column_path: i32) {
+    pub fn save_list_store(&mut self, key: String, tree_view: &TreeView, column_path: i32) {
         let mut vec_string = vec![];
         let list_store = get_list_store(tree_view);
         if let Some(iter) = list_store.iter_first() {
@@ -213,7 +214,6 @@ impl LoadSaveStruct {
         self.loaded_items.insert(key, vec_string);
     }
 
-    //noinspection RsLift
     pub fn open_save_file(&self, text_view_errors: &TextView, save_configuration: bool, manual_execution: bool) -> Option<(File, PathBuf)> {
         if let Some(proj_dirs) = ProjectDirs::from("pl", "Qarmin", "Czkawka") {
             // Lin: /home/username/.config/czkawka
@@ -260,39 +260,38 @@ impl LoadSaveStruct {
                     }
                 };
                 return Some((config_file_handler, config_file));
-            } else {
-                if !config_file.exists() || !config_file.is_file() {
-                    if manual_execution {
-                        // Don't show errors when there is no configuration file when starting app
-                        add_text_to_text_view(
-                            text_view_errors,
-                            &flg!(
-                                "saving_loading_failed_to_read_config_file",
-                                generate_translation_hashmap(vec![("path", config_file.display().to_string())])
-                            ),
-                        );
-                    }
+            }
+            if !config_file.exists() || !config_file.is_file() {
+                if manual_execution {
+                    // Don't show errors when there is no configuration file when starting app
+                    add_text_to_text_view(
+                        text_view_errors,
+                        &flg!(
+                            "saving_loading_failed_to_read_config_file",
+                            generate_translation_hashmap(vec![("path", config_file.display().to_string())])
+                        ),
+                    );
+                }
+                return None;
+            }
+
+            let config_file_handler = match File::open(&config_file) {
+                Ok(t) => t,
+                Err(e) => {
+                    add_text_to_text_view(
+                        text_view_errors,
+                        &flg!(
+                            "saving_loading_failed_to_create_config_file",
+                            generate_translation_hashmap(vec![("path", config_file.display().to_string()), ("reason", e.to_string())])
+                        ),
+                    );
                     return None;
                 }
-
-                let config_file_handler = match File::open(&config_file) {
-                    Ok(t) => t,
-                    Err(e) => {
-                        add_text_to_text_view(
-                            text_view_errors,
-                            &flg!(
-                                "saving_loading_failed_to_create_config_file",
-                                generate_translation_hashmap(vec![("path", config_file.display().to_string()), ("reason", e.to_string())])
-                            ),
-                        );
-                        return None;
-                    }
-                };
-                return Some((config_file_handler, config_file));
-            }
-        } else {
-            add_text_to_text_view(text_view_errors, flg!("saving_loading_failed_to_get_home_directory").as_str());
+            };
+            return Some((config_file_handler, config_file));
         }
+        add_text_to_text_view(text_view_errors, flg!("saving_loading_failed_to_get_home_directory").as_str());
+
         None
     }
 
@@ -310,7 +309,7 @@ impl LoadSaveStruct {
                 return;
             }
 
-            let mut header: String = "".to_string();
+            let mut header: String = String::new();
             let lines: Vec<String> = loaded_data.replace('\r', "").split('\n').map(String::from).collect::<Vec<String>>();
             for (index, line) in lines.iter().enumerate() {
                 let line = line.trim();
@@ -349,7 +348,7 @@ impl LoadSaveStruct {
         if let Some((mut config_file_handler, config_file)) = self.open_save_file(text_view_errors, true, false) {
             let mut data_saved: bool = false;
             for (key, vec_string) in &self.loaded_items {
-                match writeln!(config_file_handler, "{}", key) {
+                match writeln!(config_file_handler, "{key}") {
                     Ok(_inspected) => {
                         data_saved = true;
                     }
@@ -359,7 +358,7 @@ impl LoadSaveStruct {
                     }
                 }
                 for data in vec_string {
-                    match writeln!(config_file_handler, "{}", data) {
+                    match writeln!(config_file_handler, "{data}") {
                         Ok(_inspected) => {
                             data_saved = true;
                         }
@@ -437,6 +436,7 @@ enum LoadText {
     BrokenFilesAudio,
     BrokenFilesImage,
     BrokenFilesArchive,
+    ThreadNumber,
 }
 
 fn create_hash_map() -> (HashMap<LoadText, String>, HashMap<String, LoadText>) {
@@ -483,13 +483,14 @@ fn create_hash_map() -> (HashMap<LoadText, String>, HashMap<String, LoadText>) {
         (LoadText::BrokenFilesImage, "broken_files_image"),
         (LoadText::BrokenFilesArchive, "broken_files_archive"),
         (LoadText::GeneralIgnoreOtherFilesystems, "ignore_other_filesystems"),
+        (LoadText::ThreadNumber, "thread_number"),
     ];
     let mut hashmap_ls: HashMap<LoadText, String> = Default::default();
     let mut hashmap_sl: HashMap<String, LoadText> = Default::default();
 
     for (load_text, string) in values {
-        hashmap_ls.insert(load_text, format!("--{}", string));
-        hashmap_sl.insert(format!("--{}", string), load_text);
+        hashmap_ls.insert(load_text, format!("--{string}"));
+        hashmap_sl.insert(format!("--{string}"), load_text);
     }
 
     (hashmap_ls, hashmap_sl)
@@ -521,171 +522,175 @@ pub fn save_configuration(manual_execution: bool, upper_notebook: &GuiUpperNoteb
         &upper_notebook.tree_view_excluded_directories.clone(),
         ColumnsExcludedDirectory::Path as i32,
     );
-    saving_struct.save_var(hashmap_ls.get(&LoadText::ExcludedItems).unwrap().to_string(), upper_notebook.entry_excluded_items.text());
+    saving_struct.save_var(hashmap_ls.get(&LoadText::ExcludedItems).unwrap().to_string(), &upper_notebook.entry_excluded_items.text());
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::AllowedExtensions).unwrap().to_string(),
-        upper_notebook.entry_allowed_extensions.text(),
+        &upper_notebook.entry_allowed_extensions.text(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::MinimalFileSize).unwrap().to_string(),
-        upper_notebook.entry_general_minimal_size.text(),
+        &upper_notebook.entry_general_minimal_size.text(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::MaximalFileSize).unwrap().to_string(),
-        upper_notebook.entry_general_maximal_size.text(),
+        &upper_notebook.entry_general_maximal_size.text(),
     );
 
     // Check buttons
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::SaveAtExit).unwrap().to_string(),
-        settings.check_button_settings_save_at_exit.is_active(),
+        &settings.check_button_settings_save_at_exit.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::LoadAtStart).unwrap().to_string(),
-        settings.check_button_settings_load_at_start.is_active(),
+        &settings.check_button_settings_load_at_start.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ConfirmDeletionFiles).unwrap().to_string(),
-        settings.check_button_settings_confirm_deletion.is_active(),
+        &settings.check_button_settings_confirm_deletion.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ConfirmDeletionAllFilesInGroup).unwrap().to_string(),
-        settings.check_button_settings_confirm_group_deletion.is_active(),
+        &settings.check_button_settings_confirm_group_deletion.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ImagePreviewImage).unwrap().to_string(),
-        settings.check_button_settings_show_preview_similar_images.is_active(),
+        &settings.check_button_settings_show_preview_similar_images.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::DuplicatePreviewImage).unwrap().to_string(),
-        settings.check_button_settings_show_preview_duplicates.is_active(),
+        &settings.check_button_settings_show_preview_duplicates.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::HideHardLinks).unwrap().to_string(),
-        settings.check_button_settings_hide_hard_links.is_active(),
+        &settings.check_button_settings_hide_hard_links.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::UseCache).unwrap().to_string(),
-        settings.check_button_settings_use_cache.is_active(),
+        &settings.check_button_settings_use_cache.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::UseJsonCacheFile).unwrap().to_string(),
-        settings.check_button_settings_save_also_json.is_active(),
+        &settings.check_button_settings_save_also_json.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::DeleteToTrash).unwrap().to_string(),
-        settings.check_button_settings_use_trash.is_active(),
+        &settings.check_button_settings_use_trash.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ImageDeleteOutdatedCacheEntries).unwrap().to_string(),
-        settings.check_button_settings_similar_images_delete_outdated_cache.is_active(),
+        &settings.check_button_settings_similar_images_delete_outdated_cache.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::DuplicateDeleteOutdatedCacheEntries).unwrap().to_string(),
-        settings.check_button_settings_duplicates_delete_outdated_cache.is_active(),
+        &settings.check_button_settings_duplicates_delete_outdated_cache.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::VideoDeleteOutdatedCacheEntries).unwrap().to_string(),
-        settings.check_button_settings_similar_videos_delete_outdated_cache.is_active(),
+        &settings.check_button_settings_similar_videos_delete_outdated_cache.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::UsePrehashCache).unwrap().to_string(),
-        settings.check_button_duplicates_use_prehash_cache.is_active(),
+        &settings.check_button_duplicates_use_prehash_cache.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ShowBottomTextPanel).unwrap().to_string(),
-        settings.check_button_settings_show_text_view.is_active(),
+        &settings.check_button_settings_show_text_view.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::GeneralIgnoreOtherFilesystems).unwrap().to_string(),
-        settings.check_button_settings_one_filesystem.is_active(),
+        &settings.check_button_settings_one_filesystem.is_active(),
     );
 
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::BrokenFilesArchive).unwrap().to_string(),
-        main_notebook.check_button_broken_files_archive.is_active(),
+        &main_notebook.check_button_broken_files_archive.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::BrokenFilesImage).unwrap().to_string(),
-        main_notebook.check_button_broken_files_image.is_active(),
+        &main_notebook.check_button_broken_files_image.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::BrokenFilesAudio).unwrap().to_string(),
-        main_notebook.check_button_broken_files_audio.is_active(),
+        &main_notebook.check_button_broken_files_audio.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::BrokenFilesPdf).unwrap().to_string(),
-        main_notebook.check_button_broken_files_pdf.is_active(),
+        &main_notebook.check_button_broken_files_pdf.is_active(),
     );
 
     // Others
     saving_struct.save_var(
+        hashmap_ls.get(&LoadText::ThreadNumber).unwrap().to_string(),
+        &settings.scale_settings_number_of_threads.value().round(),
+    );
+    saving_struct.save_var(
         hashmap_ls.get(&LoadText::MinimalCacheSize).unwrap().to_string(),
-        settings.entry_settings_cache_file_minimal_size.text(),
+        &settings.entry_settings_cache_file_minimal_size.text(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::MinimalPrehashCacheSize).unwrap().to_string(),
-        settings.entry_settings_prehash_cache_file_minimal_size.text(),
+        &settings.entry_settings_prehash_cache_file_minimal_size.text(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::Language).unwrap().to_string(),
-        get_language_from_combo_box_text(settings.combo_box_settings_language.active_text().unwrap().to_string()).short_text,
+        &get_language_from_combo_box_text(&settings.combo_box_settings_language.active_text().unwrap()).short_text,
     );
 
     // Comboboxes main notebook
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ComboBoxDuplicateHashType).unwrap().to_string(),
-        main_notebook.combo_box_duplicate_hash_type.active().unwrap_or(0),
+        &main_notebook.combo_box_duplicate_hash_type.active().unwrap_or(0),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ComboBoxDuplicateCheckMethod).unwrap().to_string(),
-        main_notebook.combo_box_duplicate_check_method.active().unwrap_or(0),
+        &main_notebook.combo_box_duplicate_check_method.active().unwrap_or(0),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ComboBoxImageResizeAlgorithm).unwrap().to_string(),
-        main_notebook.combo_box_image_resize_algorithm.active().unwrap_or(0),
+        &main_notebook.combo_box_image_resize_algorithm.active().unwrap_or(0),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ComboBoxImageHashType).unwrap().to_string(),
-        main_notebook.combo_box_image_hash_algorithm.active().unwrap_or(0),
+        &main_notebook.combo_box_image_hash_algorithm.active().unwrap_or(0),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ComboBoxImageHashSize).unwrap().to_string(),
-        main_notebook.combo_box_image_hash_size.active().unwrap_or(0),
+        &main_notebook.combo_box_image_hash_size.active().unwrap_or(0),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::ComboBoxBigFiles).unwrap().to_string(),
-        main_notebook.combo_box_big_files_mode.active().unwrap_or(0),
+        &main_notebook.combo_box_big_files_mode.active().unwrap_or(0),
     );
 
     // Other2
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::DuplicateNameCaseSensitive).unwrap().to_string(),
-        main_notebook.check_button_duplicate_case_sensitive_name.is_active(),
+        &main_notebook.check_button_duplicate_case_sensitive_name.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::NumberOfBiggestFiles).unwrap().to_string(),
-        main_notebook.entry_big_files_number.text(),
+        &main_notebook.entry_big_files_number.text(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::SimilarImagesSimilarity).unwrap().to_string(),
-        main_notebook.scale_similarity_similar_images.value(),
+        &main_notebook.scale_similarity_similar_images.value(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::SimilarImagesIgnoreSameSize).unwrap().to_string(),
-        main_notebook.check_button_image_ignore_same_size.is_active(),
+        &main_notebook.check_button_image_ignore_same_size.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::SimilarVideosSimilarity).unwrap().to_string(),
-        main_notebook.scale_similarity_similar_videos.value(),
+        &main_notebook.scale_similarity_similar_videos.value(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::SimilarVideosIgnoreSameSize).unwrap().to_string(),
-        main_notebook.check_button_video_ignore_same_size.is_active(),
+        &main_notebook.check_button_video_ignore_same_size.is_active(),
     );
     saving_struct.save_var(
         hashmap_ls.get(&LoadText::MusicApproximateComparison).unwrap().to_string(),
-        main_notebook.check_button_music_approximate_comparison.is_active(),
+        &main_notebook.check_button_music_approximate_comparison.is_active(),
     );
 
     saving_struct.save_to_file(&text_view_errors);
@@ -698,7 +703,7 @@ pub fn load_configuration(
     settings: &GuiSettings,
     text_view_errors: &TextView,
     scrolled_window_errors: &ScrolledWindow,
-    arguments: Vec<OsString>,
+    arguments: &[OsString],
 ) {
     let text_view_errors = text_view_errors.clone();
 
@@ -708,7 +713,7 @@ pub fn load_configuration(
     loaded_entries.open_and_read_content(&text_view_errors, manual_execution);
 
     // Load here language, default system language could change value in settings so we don't want to lose this value
-    let short_language = get_language_from_combo_box_text(settings.combo_box_settings_language.active_text().unwrap().to_string())
+    let short_language = get_language_from_combo_box_text(&settings.combo_box_settings_language.active_text().unwrap())
         .short_text
         .to_string();
 
@@ -718,18 +723,18 @@ pub fn load_configuration(
     // Loading data from hashmaps
     let (hashmap_ls, _hashmap_sl) = create_hash_map();
 
-    let mut included_directories: Vec<String> = loaded_entries.get_vector_string(hashmap_ls.get(&LoadText::IncludedDirectories).unwrap().clone(), included_directories);
-    let mut excluded_directories: Vec<String> = loaded_entries.get_vector_string(hashmap_ls.get(&LoadText::ExcludedDirectories).unwrap().clone(), excluded_directories);
+    let mut included_directories: Vec<String> = loaded_entries.get_vector_string(hashmap_ls.get(&LoadText::IncludedDirectories).unwrap(), included_directories);
+    let mut excluded_directories: Vec<String> = loaded_entries.get_vector_string(hashmap_ls.get(&LoadText::ExcludedDirectories).unwrap(), excluded_directories);
     let excluded_items: String = loaded_entries.get_string(
         hashmap_ls.get(&LoadText::ExcludedItems).unwrap().clone(),
         upper_notebook.entry_excluded_items.text().to_string(),
     );
-    let allowed_extensions: String = loaded_entries.get_string(hashmap_ls.get(&LoadText::AllowedExtensions).unwrap().clone(), "".to_string());
+    let allowed_extensions: String = loaded_entries.get_string(hashmap_ls.get(&LoadText::AllowedExtensions).unwrap().clone(), String::new());
     let minimal_file_size: String = loaded_entries.get_integer_string(hashmap_ls.get(&LoadText::MinimalFileSize).unwrap().clone(), DEFAULT_MINIMAL_FILE_SIZE.to_string());
     let maximal_file_size: String = loaded_entries.get_integer_string(hashmap_ls.get(&LoadText::MaximalFileSize).unwrap().clone(), DEFAULT_MAXIMAL_FILE_SIZE.to_string());
 
     let loading_at_start: bool = loaded_entries.get_bool(hashmap_ls.get(&LoadText::LoadAtStart).unwrap().clone(), DEFAULT_LOAD_AT_START);
-    let saving_at_exit: bool = loaded_entries.get_bool(hashmap_ls.get(&LoadText::SaveAtExit).unwrap().clone(), DEFAULT_SAVE_ON_EXIT);
+    let mut saving_at_exit: bool = loaded_entries.get_bool(hashmap_ls.get(&LoadText::SaveAtExit).unwrap().clone(), DEFAULT_SAVE_ON_EXIT);
     let confirm_deletion: bool = loaded_entries.get_bool(hashmap_ls.get(&LoadText::ConfirmDeletionFiles).unwrap().clone(), DEFAULT_CONFIRM_DELETION);
     let confirm_group_deletion: bool = loaded_entries.get_bool(hashmap_ls.get(&LoadText::ConfirmDeletionAllFilesInGroup).unwrap().clone(), DEFAULT_CONFIRM_GROUP_DELETION);
     let show_previews_similar_images: bool = loaded_entries.get_bool(hashmap_ls.get(&LoadText::ImagePreviewImage).unwrap().clone(), DEFAULT_SHOW_IMAGE_PREVIEW);
@@ -791,76 +796,74 @@ pub fn load_configuration(
     );
 
     let check_button_broken_files_archive = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesArchive).unwrap().clone(), DEFAULT_BROKEN_FILES_ARCHIVE);
-    let check_button_broken_files_pdf = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesPdf).unwrap().clone(), DEFAULT_BROKEN_FILES_ARCHIVE);
-    let check_button_broken_files_image = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesImage).unwrap().clone(), DEFAULT_BROKEN_FILES_ARCHIVE);
-    let check_button_broken_files_audio = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesAudio).unwrap().clone(), DEFAULT_BROKEN_FILES_ARCHIVE);
+    let check_button_broken_files_pdf = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesPdf).unwrap().clone(), DEFAULT_BROKEN_FILES_PDF);
+    let check_button_broken_files_image = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesImage).unwrap().clone(), DEFAULT_BROKEN_FILES_IMAGE);
+    let check_button_broken_files_audio = loaded_entries.get_object(hashmap_ls.get(&LoadText::BrokenFilesAudio).unwrap().clone(), DEFAULT_BROKEN_FILES_AUDIO);
+    let thread_number = loaded_entries.get_object(hashmap_ls.get(&LoadText::ThreadNumber).unwrap().clone(), DEFAULT_THREAD_NUMBER);
 
-    // Setting data
-    if manual_execution || loading_at_start {
-        {
-            // Handle here arguments that were added to app e.g. czkawka_gui /home --/home/roman
-            if loading_at_start && arguments.len() > 1 {
-                let iter_i = arguments.iter().skip(1);
-                let iter_e = iter_i.clone();
-                included_directories = iter_i
-                    .filter_map(|e| {
-                        let r = e.to_string_lossy().to_string();
-                        if !r.starts_with("--") {
-                            let path = Path::new(&r);
-                            if !path.exists() {
-                                return None;
-                            }
-                            match path.canonicalize() {
-                                Ok(r) => Some(r.to_string_lossy().to_string()),
-                                Err(_) => None,
-                            }
-                        } else {
-                            None
+    let mut set_start_folders = false;
+    if !manual_execution {
+        // Handle here arguments that were added to app e.g. czkawka_gui /home --/home/roman
+        if arguments.len() > 1 {
+            let iter_i = arguments.iter().skip(1);
+            let iter_e = iter_i.clone();
+            let inc_dir = iter_i
+                .filter_map(|e| {
+                    let r = e.to_string_lossy().to_string();
+                    if !r.starts_with("--") {
+                        let path = Path::new(&r);
+                        if !path.exists() {
+                            return None;
                         }
-                    })
-                    .collect::<Vec<_>>();
-                excluded_directories = iter_e
-                    .filter_map(|e| {
-                        let r = e.to_string_lossy().to_string();
-                        if let Some(r) = r.strip_prefix("--") {
-                            let path = Path::new(&r);
-                            if !path.exists() {
-                                return None;
-                            }
-                            match path.canonicalize() {
-                                Ok(r) => Some(r.to_string_lossy().to_string()),
-                                Err(_) => None,
-                            }
-                        } else {
-                            None
+                        match path.canonicalize() {
+                            Ok(r) => Some(r.to_string_lossy().to_string()),
+                            Err(_) => None,
                         }
-                    })
-                    .collect::<Vec<_>>();
-            }
+                    } else {
+                        None
+                    }
+                })
+                .collect::<Vec<_>>();
+            let exc_dir = iter_e
+                .filter_map(|e| {
+                    let r = e.to_string_lossy().to_string();
+                    if let Some(r) = r.strip_prefix("--") {
+                        let path = Path::new(&r);
+                        if !path.exists() {
+                            return None;
+                        }
+                        match path.canonicalize() {
+                            Ok(r) => Some(r.to_string_lossy().to_string()),
+                            Err(_) => None,
+                        }
+                    } else {
+                        None
+                    }
+                })
+                .collect::<Vec<_>>();
 
-            // Include Directories
-            let tree_view_included_directories = upper_notebook.tree_view_included_directories.clone();
-            let list_store = get_list_store(&tree_view_included_directories);
-            list_store.clear();
-
-            for directory in included_directories {
-                let values: [(u32, &dyn ToValue); 2] = [
-                    (ColumnsIncludedDirectory::Path as u32, &directory),
-                    (ColumnsIncludedDirectory::ReferenceButton as u32, &false),
-                ];
-                list_store.set(&list_store.append(), &values);
-            }
-
-            //// Exclude Directories
-            let tree_view_excluded_directories = upper_notebook.tree_view_excluded_directories.clone();
-            let list_store = get_list_store(&tree_view_excluded_directories);
-            list_store.clear();
-
-            for directory in excluded_directories {
-                let values: [(u32, &dyn ToValue); 1] = [(ColumnsExcludedDirectory::Path as u32, &directory)];
-                list_store.set(&list_store.append(), &values);
+            if inc_dir.is_empty() {
+                println!("Arguments {arguments:?} should contains at least one directory to include");
+            } else {
+                included_directories = inc_dir;
+                excluded_directories = exc_dir;
+                saving_at_exit = false;
+                set_start_folders = true;
             }
         }
+    }
+
+    if manual_execution || loading_at_start || set_start_folders {
+        set_directories(
+            &upper_notebook.tree_view_included_directories,
+            &upper_notebook.tree_view_excluded_directories,
+            &included_directories,
+            &excluded_directories,
+        );
+    }
+
+    // Setting data
+    if loading_at_start || manual_execution {
         //// Language ComboBoxText
         {
             for (index, lang) in LANGUAGES_ALL.iter().enumerate() {
@@ -947,10 +950,38 @@ pub fn load_configuration(
 
         main_notebook.scale_similarity_similar_images.set_range(0_f64, SIMILAR_VALUES[index][5] as f64);
         main_notebook.scale_similarity_similar_images.set_fill_level(SIMILAR_VALUES[index][5] as f64);
-
+        main_notebook.scale_similarity_similar_images.connect_change_value(scale_step_function);
         main_notebook.scale_similarity_similar_images.set_value(similar_images_similarity as f64);
+
+        settings.scale_settings_number_of_threads.set_range(0_f64, get_default_number_of_threads() as f64);
+        settings.scale_settings_number_of_threads.set_fill_level(get_default_number_of_threads() as f64);
+        settings.scale_settings_number_of_threads.connect_change_value(scale_step_function);
+        settings.scale_settings_number_of_threads.set_value(thread_number as f64);
     } else {
         settings.check_button_settings_load_at_start.set_active(false);
+    }
+}
+
+fn set_directories(tree_view_included_directories: &TreeView, tree_view_excluded_directories: &TreeView, included_directories: &[String], excluded_directories: &[String]) {
+    // Include Directories
+    let list_store = get_list_store(tree_view_included_directories);
+    list_store.clear();
+
+    for directory in included_directories {
+        let values: [(u32, &dyn ToValue); 2] = [
+            (ColumnsIncludedDirectory::Path as u32, &directory),
+            (ColumnsIncludedDirectory::ReferenceButton as u32, &false),
+        ];
+        list_store.set(&list_store.append(), &values);
+    }
+
+    //// Exclude Directories
+    let list_store = get_list_store(tree_view_excluded_directories);
+    list_store.clear();
+
+    for directory in excluded_directories {
+        let values: [(u32, &dyn ToValue); 1] = [(ColumnsExcludedDirectory::Path as u32, &directory)];
+        list_store.set(&list_store.append(), &values);
     }
 }
 
@@ -985,7 +1016,7 @@ pub fn reset_configuration(manual_clearing: bool, upper_notebook: &GuiUpperNoteb
                     add_text_to_text_view(&text_view_errors, "Failed to read current directory, setting C:\\ instead");
                     "C:\\".to_string()
                 } else {
-                    "".to_string()
+                    String::new()
                 }
             }
         };
