@@ -98,10 +98,8 @@ fn popover_all_except_oldest_newest(
             let mut tree_iter_array: Vec<TreeIter> = Vec::new();
             let mut used_index: Option<usize> = None;
             let mut current_index: usize = 0;
-            let mut modification_time_min_max: u64 = match except_oldest {
-                true => u64::MAX,
-                false => 0,
-            };
+
+            let mut modification_time_min_max: u64 = if except_oldest { u64::MAX } else { 0 };
 
             let mut file_length: usize = 0;
 
@@ -135,7 +133,7 @@ fn popover_all_except_oldest_newest(
                     break;
                 }
             }
-            if used_index == None {
+            if used_index.is_none() {
                 continue;
             }
             for (index, tree_iter) in tree_iter_array.iter().enumerate() {
@@ -172,10 +170,7 @@ fn popover_one_oldest_newest(
             let mut tree_iter_array: Vec<TreeIter> = Vec::new();
             let mut used_index: Option<usize> = None;
             let mut current_index: usize = 0;
-            let mut modification_time_min_max: u64 = match check_oldest {
-                true => u64::MAX,
-                false => 0,
-            };
+            let mut modification_time_min_max: u64 = if check_oldest { u64::MAX } else { 0 };
 
             let mut file_length: usize = 0;
 
@@ -210,7 +205,7 @@ fn popover_one_oldest_newest(
                     break;
                 }
             }
-            if used_index == None {
+            if used_index.is_none() {
                 continue;
             }
             for (index, tree_iter) in tree_iter_array.iter().enumerate() {
@@ -242,55 +237,59 @@ fn popover_custom_select_unselect(
 ) {
     popover.popdown();
 
-    let window_title = match select_things {
-        false => flg!("popover_custom_mode_unselect"),
-        true => flg!("popover_custom_mode_select"),
+    let window_title = if select_things {
+        flg!("popover_custom_mode_select")
+    } else {
+        flg!("popover_custom_mode_unselect")
     };
 
     // Dialog for select/unselect items
     {
-        let dialog = gtk4::Dialog::builder().title(&window_title).transient_for(window_main).modal(true).build();
+        let dialog = gtk4::Dialog::builder().title(window_title).transient_for(window_main).modal(true).build();
         dialog.add_button(&flg!("general_ok_button"), ResponseType::Ok);
         dialog.add_button(&flg!("general_close_button"), ResponseType::Cancel);
 
-        let check_button_path = gtk4::CheckButton::builder().label(&flg!("popover_custom_regex_path_label")).build();
-        let check_button_name = gtk4::CheckButton::builder().label(&flg!("popover_custom_regex_name_label")).build();
-        let check_button_rust_regex = gtk4::CheckButton::builder().label(&flg!("popover_custom_regex_regex_label")).build();
+        let check_button_path = gtk4::CheckButton::builder()
+            .label(flg!("popover_custom_regex_path_label"))
+            .tooltip_text(flg!("popover_custom_path_check_button_entry_tooltip"))
+            .build();
+        let check_button_name = gtk4::CheckButton::builder()
+            .label(flg!("popover_custom_regex_name_label"))
+            .tooltip_text(flg!("popover_custom_name_check_button_entry_tooltip"))
+            .build();
+        let check_button_rust_regex = gtk4::CheckButton::builder()
+            .label(flg!("popover_custom_regex_regex_label"))
+            .tooltip_text(flg!("popover_custom_regex_check_button_entry_tooltip"))
+            .build();
 
-        let check_button_case_sensitive = gtk4::CheckButton::builder().label(&flg!("popover_custom_case_sensitive_check_button")).build();
-        check_button_case_sensitive.set_active(false);
+        let check_button_case_sensitive = gtk4::CheckButton::builder()
+            .label(flg!("popover_custom_case_sensitive_check_button"))
+            .tooltip_text(flg!("popover_custom_case_sensitive_check_button_tooltip"))
+            .active(false)
+            .build();
 
-        let check_button_select_not_all_results = gtk4::CheckButton::builder().label(&flg!("popover_custom_all_in_group_label")).build();
-        check_button_select_not_all_results.set_active(true);
+        let check_button_select_not_all_results = gtk4::CheckButton::builder()
+            .label(flg!("popover_custom_all_in_group_label"))
+            .tooltip_text(flg!("popover_custom_not_all_check_button_tooltip"))
+            .active(true)
+            .build();
 
-        let entry_path = gtk4::Entry::new();
-        let entry_name = gtk4::Entry::new();
-        let entry_rust_regex = gtk4::Entry::new();
-        entry_rust_regex.set_sensitive(false); // By default check button regex is disabled
+        let entry_path = gtk4::Entry::builder().tooltip_text(flg!("popover_custom_path_check_button_entry_tooltip")).build();
+        let entry_name = gtk4::Entry::builder().tooltip_text(flg!("popover_custom_name_check_button_entry_tooltip")).build();
+        let entry_rust_regex = gtk4::Entry::builder()
+            .tooltip_text(flg!("popover_custom_regex_check_button_entry_tooltip"))
+            .sensitive(false)
+            .build(); // By default check button regex is disabled
 
         let label_regex_valid = gtk4::Label::new(None);
 
-        // Tooltips
-        {
-            check_button_path.set_tooltip_text(Some(&flg!("popover_custom_path_check_button_entry_tooltip")));
-            entry_path.set_tooltip_text(Some(&flg!("popover_custom_path_check_button_entry_tooltip")));
-
-            check_button_name.set_tooltip_text(Some(&flg!("popover_custom_name_check_button_entry_tooltip")));
-            entry_name.set_tooltip_text(Some(&flg!("popover_custom_name_check_button_entry_tooltip")));
-
-            check_button_rust_regex.set_tooltip_text(Some(&flg!("popover_custom_regex_check_button_entry_tooltip")));
-            entry_rust_regex.set_tooltip_text(Some(&flg!("popover_custom_regex_check_button_entry_tooltip")));
-
-            check_button_case_sensitive.set_tooltip_text(Some(&flg!("popover_custom_case_sensitive_check_button_tooltip")));
-            check_button_select_not_all_results.set_tooltip_text(Some(&flg!("popover_custom_not_all_check_button_tooltip")));
-        }
         {
             let label_regex_valid = label_regex_valid.clone();
             entry_rust_regex.connect_changed(move |entry_rust_regex| {
                 let message;
                 let text_to_check = entry_rust_regex.text().to_string();
                 if text_to_check.is_empty() {
-                    message = "".to_string();
+                    message = String::new();
                 } else {
                     match Regex::new(&text_to_check) {
                         Ok(_) => message = flg!("popover_valid_regex"),
@@ -336,9 +335,7 @@ fn popover_custom_select_unselect(
         // Configure look of things
         {
             // TODO Label should have const width, and rest should fill entry, but for now is 50%-50%
-            let grid = gtk4::Grid::new();
-            grid.set_row_homogeneous(true);
-            grid.set_column_homogeneous(true);
+            let grid = gtk4::Grid::builder().row_homogeneous(true).column_homogeneous(true).build();
 
             grid.attach(&check_button_name, 0, 1, 1, 1);
             grid.attach(&check_button_path, 0, 2, 1, 1);
@@ -373,7 +370,7 @@ fn popover_custom_select_unselect(
             #[cfg(target_family = "windows")]
             let path_wildcard = path_wildcard.replace("/", "\\");
 
-            if response_type == gtk4::ResponseType::Ok {
+            if response_type == ResponseType::Ok {
                 let check_path = check_button_path.is_active();
                 let check_name = check_button_name.is_active();
                 let check_regex = check_button_rust_regex.is_active();
@@ -382,26 +379,25 @@ fn popover_custom_select_unselect(
                 let check_all_selected = check_button_select_not_all_results.is_active();
 
                 if check_button_path.is_active() || check_button_name.is_active() || check_button_rust_regex.is_active() {
-                    let compiled_regex = match check_regex {
-                        true => match Regex::new(&regex_wildcard) {
-                            Ok(t) => t,
-                            Err(_) => {
-                                eprintln!("What? Regex should compile properly.");
-                                confirmation_dialog_select_unselect.close();
-                                return;
-                            }
-                        },
-                        false => Regex::new("").unwrap(),
+                    let compiled_regex = if check_regex {
+                        if let Ok(t) = Regex::new(&regex_wildcard) {
+                            t
+                        } else {
+                            eprintln!("What? Regex should compile properly.");
+                            confirmation_dialog_select_unselect.close();
+                            return;
+                        }
+                    } else {
+                        Regex::new("").unwrap()
                     };
 
                     let model = get_list_store(&tree_view);
 
-                    let iter = match model.iter_first() {
-                        Some(t) => t,
-                        None => {
-                            confirmation_dialog_select_unselect.close();
-                            return;
-                        }
+                    let iter = if let Some(t) = model.iter_first() {
+                        t
+                    } else {
+                        confirmation_dialog_select_unselect.close();
+                        return;
                     };
 
                     let mut number_of_all_things = 0;
@@ -452,7 +448,7 @@ fn popover_custom_select_unselect(
                                         need_to_change_thing = true;
                                     }
                                 } else {
-                                    if Common::regex_check(&name_wildcard.to_lowercase(), &name.to_lowercase()) {
+                                    if Common::regex_check(&name_wildcard.to_lowercase(), name.to_lowercase()) {
                                         need_to_change_thing = true;
                                     }
                                 }
@@ -463,7 +459,7 @@ fn popover_custom_select_unselect(
                                         need_to_change_thing = true;
                                     }
                                 } else {
-                                    if Common::regex_check(&path_wildcard.to_lowercase(), &path.to_lowercase()) {
+                                    if Common::regex_check(&path_wildcard.to_lowercase(), path.to_lowercase()) {
                                         need_to_change_thing = true;
                                     }
                                 }
@@ -524,14 +520,8 @@ fn popover_all_except_biggest_smallest(
             let mut tree_iter_array: Vec<TreeIter> = Vec::new();
             let mut used_index: Option<usize> = None;
             let mut current_index: usize = 0;
-            let mut size_as_bytes_min_max: u64 = match except_biggest {
-                true => 0,
-                false => u64::MAX,
-            };
-            let mut number_of_pixels_min_max: u64 = match except_biggest {
-                true => 0,
-                false => u64::MAX,
-            };
+            let mut size_as_bytes_min_max: u64 = if except_biggest { 0 } else { u64::MAX };
+            let mut number_of_pixels_min_max: u64 = if except_biggest { 0 } else { u64::MAX };
 
             loop {
                 if model.get::<bool>(&iter, column_header) {
@@ -547,7 +537,7 @@ fn popover_all_except_biggest_smallest(
                 if let Some(column_dimensions) = column_dimensions {
                     let dimensions_string = model.get::<String>(&iter, column_dimensions);
 
-                    let dimensions = change_dimension_to_krotka(dimensions_string);
+                    let dimensions = change_dimension_to_krotka(&dimensions_string);
                     let number_of_pixels = dimensions.0 * dimensions.1;
 
                     if except_biggest {
@@ -584,7 +574,7 @@ fn popover_all_except_biggest_smallest(
                     break;
                 }
             }
-            if used_index == None {
+            if used_index.is_none() {
                 continue;
             }
             for (index, tree_iter) in tree_iter_array.iter().enumerate() {
@@ -604,9 +594,9 @@ fn popover_all_except_biggest_smallest(
     popover.popdown();
 }
 
-pub fn connect_popovers(gui_data: &GuiData) {
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_all = gui_data.popovers.buttons_popover_select_all.clone();
+pub fn connect_popover_select(gui_data: &GuiData) {
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_all = gui_data.popovers_select.buttons_popover_select_all.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
 
@@ -618,8 +608,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         popover_select_all(&popover_select, tree_view, nb_object.column_selection as u32, nb_object.column_header);
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_unselect_all = gui_data.popovers.buttons_popover_unselect_all.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_unselect_all = gui_data.popovers_select.buttons_popover_unselect_all.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_unselect_all.connect_clicked(move |_| {
@@ -630,8 +620,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         popover_unselect_all(&popover_select, tree_view, nb_object.column_selection as u32);
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_reverse = gui_data.popovers.buttons_popover_reverse.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_reverse = gui_data.popovers_select.buttons_popover_reverse.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_reverse.connect_clicked(move |_| {
@@ -642,8 +632,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         popover_reverse(&popover_select, tree_view, nb_object.column_selection as u32, nb_object.column_header);
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_all_except_oldest = gui_data.popovers.buttons_popover_select_all_except_oldest.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_all_except_oldest = gui_data.popovers_select.buttons_popover_select_all_except_oldest.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_select_all_except_oldest.connect_clicked(move |_| {
@@ -662,8 +652,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_all_except_newest = gui_data.popovers.buttons_popover_select_all_except_newest.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_all_except_newest = gui_data.popovers_select.buttons_popover_select_all_except_newest.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_select_all_except_newest.connect_clicked(move |_| {
@@ -682,8 +672,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_one_oldest = gui_data.popovers.buttons_popover_select_one_oldest.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_one_oldest = gui_data.popovers_select.buttons_popover_select_one_oldest.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_select_one_oldest.connect_clicked(move |_| {
@@ -702,8 +692,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_one_newest = gui_data.popovers.buttons_popover_select_one_newest.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_one_newest = gui_data.popovers_select.buttons_popover_select_one_newest.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_select_one_newest.connect_clicked(move |_| {
@@ -722,8 +712,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_custom = gui_data.popovers.buttons_popover_select_custom.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_custom = gui_data.popovers_select.buttons_popover_select_custom.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     let window_main = gui_data.window_main.clone();
@@ -744,8 +734,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_unselect_custom = gui_data.popovers.buttons_popover_unselect_custom.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_unselect_custom = gui_data.popovers_select.buttons_popover_unselect_custom.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     let window_main = gui_data.window_main.clone();
@@ -766,8 +756,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_all_images_except_biggest = gui_data.popovers.buttons_popover_select_all_images_except_biggest.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_all_images_except_biggest = gui_data.popovers_select.buttons_popover_select_all_images_except_biggest.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_select_all_images_except_biggest.connect_clicked(move |_| {
@@ -786,8 +776,8 @@ pub fn connect_popovers(gui_data: &GuiData) {
         );
     });
 
-    let popover_select = gui_data.popovers.popover_select.clone();
-    let buttons_popover_select_all_images_except_smallest = gui_data.popovers.buttons_popover_select_all_images_except_smallest.clone();
+    let popover_select = gui_data.popovers_select.popover_select.clone();
+    let buttons_popover_select_all_images_except_smallest = gui_data.popovers_select.buttons_popover_select_all_images_except_smallest.clone();
     let notebook_main = gui_data.main_notebook.notebook_main.clone();
     let main_tree_views = gui_data.main_notebook.get_main_tree_views();
     buttons_popover_select_all_images_except_smallest.connect_clicked(move |_| {
