@@ -6,12 +6,14 @@ pub mod connect_button_move;
 pub mod connect_button_save;
 pub mod connect_button_search;
 pub mod connect_button_select;
+pub mod connect_button_sort;
 pub mod connect_button_stop;
 pub mod connect_change_language;
 pub mod connect_duplicate_buttons;
 pub mod connect_header_buttons;
 pub mod connect_notebook_tabs;
-pub mod connect_popovers;
+pub mod connect_popovers_select;
+pub mod connect_popovers_sort;
 pub mod connect_progress_window;
 pub mod connect_selection_of_directories;
 pub mod connect_settings;
