@@ -153,50 +153,50 @@ impl GuiMainNotebook {
         let tree_view_bad_extensions: TreeView = TreeView::new();
 
         let evk_tree_view_duplicate_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_duplicate_finder.add_controller(&evk_tree_view_duplicate_finder);
+        tree_view_duplicate_finder.add_controller(evk_tree_view_duplicate_finder.clone());
         let evk_tree_view_empty_folder_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_empty_folder_finder.add_controller(&evk_tree_view_empty_folder_finder);
+        tree_view_empty_folder_finder.add_controller(evk_tree_view_empty_folder_finder.clone());
         let evk_tree_view_empty_files_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_empty_files_finder.add_controller(&evk_tree_view_empty_files_finder);
+        tree_view_empty_files_finder.add_controller(evk_tree_view_empty_files_finder.clone());
         let evk_tree_view_temporary_files_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_temporary_files_finder.add_controller(&evk_tree_view_temporary_files_finder);
+        tree_view_temporary_files_finder.add_controller(evk_tree_view_temporary_files_finder.clone());
         let evk_tree_view_big_files_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_big_files_finder.add_controller(&evk_tree_view_big_files_finder);
+        tree_view_big_files_finder.add_controller(evk_tree_view_big_files_finder.clone());
         let evk_tree_view_similar_images_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_similar_images_finder.add_controller(&evk_tree_view_similar_images_finder);
+        tree_view_similar_images_finder.add_controller(evk_tree_view_similar_images_finder.clone());
         let evk_tree_view_similar_videos_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_similar_videos_finder.add_controller(&evk_tree_view_similar_videos_finder);
+        tree_view_similar_videos_finder.add_controller(evk_tree_view_similar_videos_finder.clone());
         let evk_tree_view_same_music_finder: EventControllerKey = EventControllerKey::new();
-        tree_view_same_music_finder.add_controller(&evk_tree_view_same_music_finder);
+        tree_view_same_music_finder.add_controller(evk_tree_view_same_music_finder.clone());
         let evk_tree_view_invalid_symlinks: EventControllerKey = EventControllerKey::new();
-        tree_view_invalid_symlinks.add_controller(&evk_tree_view_invalid_symlinks);
+        tree_view_invalid_symlinks.add_controller(evk_tree_view_invalid_symlinks.clone());
         let evk_tree_view_broken_files: EventControllerKey = EventControllerKey::new();
-        tree_view_broken_files.add_controller(&evk_tree_view_broken_files);
+        tree_view_broken_files.add_controller(evk_tree_view_broken_files.clone());
         let evk_tree_view_bad_extensions: EventControllerKey = EventControllerKey::new();
-        tree_view_bad_extensions.add_controller(&evk_tree_view_bad_extensions);
+        tree_view_bad_extensions.add_controller(evk_tree_view_bad_extensions.clone());
 
         let gc_tree_view_duplicate_finder: GestureClick = GestureClick::new();
-        tree_view_duplicate_finder.add_controller(&gc_tree_view_duplicate_finder);
+        tree_view_duplicate_finder.add_controller(gc_tree_view_duplicate_finder.clone());
         let gc_tree_view_empty_folder_finder: GestureClick = GestureClick::new();
-        tree_view_empty_folder_finder.add_controller(&gc_tree_view_empty_folder_finder);
+        tree_view_empty_folder_finder.add_controller(gc_tree_view_empty_folder_finder.clone());
         let gc_tree_view_empty_files_finder: GestureClick = GestureClick::new();
-        tree_view_empty_files_finder.add_controller(&gc_tree_view_empty_files_finder);
+        tree_view_empty_files_finder.add_controller(gc_tree_view_empty_files_finder.clone());
         let gc_tree_view_temporary_files_finder: GestureClick = GestureClick::new();
-        tree_view_temporary_files_finder.add_controller(&gc_tree_view_temporary_files_finder);
+        tree_view_temporary_files_finder.add_controller(gc_tree_view_temporary_files_finder.clone());
         let gc_tree_view_big_files_finder: GestureClick = GestureClick::new();
-        tree_view_big_files_finder.add_controller(&gc_tree_view_big_files_finder);
+        tree_view_big_files_finder.add_controller(gc_tree_view_big_files_finder.clone());
         let gc_tree_view_similar_images_finder: GestureClick = GestureClick::new();
-        tree_view_similar_images_finder.add_controller(&gc_tree_view_similar_images_finder);
+        tree_view_similar_images_finder.add_controller(gc_tree_view_similar_images_finder.clone());
         let gc_tree_view_similar_videos_finder: GestureClick = GestureClick::new();
-        tree_view_similar_videos_finder.add_controller(&gc_tree_view_similar_videos_finder);
+        tree_view_similar_videos_finder.add_controller(gc_tree_view_similar_videos_finder.clone());
         let gc_tree_view_same_music_finder: GestureClick = GestureClick::new();
-        tree_view_same_music_finder.add_controller(&gc_tree_view_same_music_finder);
+        tree_view_same_music_finder.add_controller(gc_tree_view_same_music_finder.clone());
         let gc_tree_view_invalid_symlinks: GestureClick = GestureClick::new();
-        tree_view_invalid_symlinks.add_controller(&gc_tree_view_invalid_symlinks);
+        tree_view_invalid_symlinks.add_controller(gc_tree_view_invalid_symlinks.clone());
         let gc_tree_view_broken_files: GestureClick = GestureClick::new();
-        tree_view_broken_files.add_controller(&gc_tree_view_broken_files);
+        tree_view_broken_files.add_controller(gc_tree_view_broken_files.clone());
         let gc_tree_view_bad_extensions: GestureClick = GestureClick::new();
-        tree_view_bad_extensions.add_controller(&gc_tree_view_bad_extensions);
+        tree_view_bad_extensions.add_controller(gc_tree_view_bad_extensions.clone());
 
         let combo_box_duplicate_check_method: ComboBoxText = builder.object("combo_box_duplicate_check_method").unwrap();
         let combo_box_duplicate_hash_type: ComboBoxText = builder.object("combo_box_duplicate_hash_type").unwrap();
