@@ -4,7 +4,8 @@ mod gui_compare_images;
 pub mod gui_data;
 mod gui_header;
 pub mod gui_main_notebook;
-pub mod gui_popovers;
+pub mod gui_popovers_select;
+pub mod gui_popovers_sort;
 mod gui_progress_dialog;
 pub mod gui_settings;
 pub mod gui_upper_notebook;
